integrate parameter parsing into dispatcher
