--- ts/packages/agentSdk/jest.config.cjs
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-module.exports = require("../../jest.config.js");
