+++ ts/packages/agentSdk/src/helpers/commandHelpers.ts
@@ -7,12 +7,10 @@ import {
     CommandDescriptor,
     CommandDescriptors,
     CommandDescriptorTable,
-    ParameterDefinitions,
 } from "../command.js";
-import { parseParams, ParsedCommandParams } from "./parameterHelpers.js";
+import { ParameterDefinitions, ParsedCommandParams } from "../parameters.js";
 
 export {
-    ParsedCommandParams,
     resolveFlag,
     getFlagMultiple,
     getFlagType,
@@ -65,7 +63,7 @@ export function getCommandInterface(
         getCommands: async () => handlers,
         executeCommand: async (
             commands: string[],
-            args: string,
+            params: ParsedCommandParams<ParameterDefinitions> | undefined,
             context: ActionContext<unknown>,
             attachments?: string[],
         ) => {
@@ -99,19 +97,22 @@ export function getCommandInterface(
                 }
                 curr = curr.defaultSubCommand;
             }
+
             if (isCommandHandlerNoParams(curr)) {
-                if (args.trim() !== "") {
+                if (params !== undefined) {
                     throw new Error(
-                        `No parameters expected for command '@${commandPrefix.join(" ")}'`,
+                        `Command '@${commandPrefix.join(" ")}' does not accept parameters`,
                     );
                 }
                 await curr.run(context, undefined, attachments);
+                return;
             } else {
-                await curr.run(
-                    context,
-                    parseParams(args, curr.parameters),
-                    attachments,
-                );
+                if (params === undefined) {
+                    throw new Error(
+                        `Command '@${commandPrefix.join(" ")}' expects parameters`,
+                    );
+                }
+                await curr.run(context, params, attachments);
             }
         },
     };
