--- ts/packages/agentSdk/test/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": ".",
-    "outDir": "../dist/test",
-    "types": ["node", "jest"],
-  },
-  "include": ["./**/*"],
-  "ts-node": {
-    "esm": true
-  },
-  "references": [
-    { "path": "../src" }
-  ]
-}
