+++ ts/packages/agentSdk/tsconfig.json
@@ -5,8 +5,7 @@
   },
   "include": [],
   "references": [
-    { "path": "./src" },
-    { "path": "./test" }
+    { "path": "./src" }
   ],
   "ts-node": {
     "esm": true
